+++ src/index.ts
@@ -1,0 +1,2 @@
+export { Release } from './lib/release/release';
+export { confirmVersion } from './lib/confirm/confirm';